Moves examples into magnolia package
