+++ src/examples/magnolia/examples/JavaAnnotatedCase.scala
