+++ src/examples/magnolia/examples/print.scala
